+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Removed `EnforceStrength` in login options as it has changed from string to bool which was causing a parsing error.
 * Added common request options (only supports DryRun for now)
 * Added DryRunResponse option to return a fake response containing the Request that would have been sent
+* Added `UnsilenceLogger` to re-enable logger output after using `SilenceLogger`
 
 ## Released